Refacotr code
