+++ components/org.wso2.carbon.crypto.provider/src/main/java/org/wso2/carbon/crypto/provider/KeyStoreBasedInternalCryptoProvider.java
@@ -18,21 +18,28 @@
 
 package org.wso2.carbon.crypto.provider;
 
+import com.google.gson.Gson;
+import org.apache.axiom.om.util.Base64;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.wso2.carbon.crypto.api.CipherMetaDataHolder;
 import org.wso2.carbon.crypto.api.CryptoException;
 import org.wso2.carbon.crypto.api.InternalCryptoProvider;
 
+import java.nio.charset.Charset;
 import java.security.InvalidKeyException;
 import java.security.Key;
 import java.security.KeyStore;
 import java.security.KeyStoreException;
+import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
 import java.security.PrivateKey;
 import java.security.UnrecoverableKeyException;
 import java.security.cert.Certificate;
+import java.security.cert.CertificateEncodingException;
+
 import javax.crypto.BadPaddingException;
 import javax.crypto.Cipher;
 import javax.crypto.IllegalBlockSizeException;
@@ -45,6 +52,7 @@ public class KeyStoreBasedInternalCryptoProvider implements InternalCryptoProvid
 
     private static Log log = LogFactory.getLog(KeyStoreBasedInternalCryptoProvider.class);
     private static final String DEFAULT_ASSYMETRIC_CRYPTO_ALGORITHM = "RSA";
+    private static final String CIPHER_TRANSFORMATION_SYSTEM_PROPERTY = "org.wso2.CipherTransformation";
 
     private KeyStore keyStore;
     private String keyAlias;
@@ -156,6 +164,33 @@ public class KeyStoreBasedInternalCryptoProvider implements InternalCryptoProvid
         }
     }
 
+    @Override
+    public byte[] encrypt(byte[] cleartext, String algorithm, String javaSecurityAPIProvider,
+                          boolean returnSelfContainedCipherText) throws CryptoException {
+
+        byte[] encryptedKey = encrypt(cleartext, algorithm, javaSecurityAPIProvider);
+        String cipherTransformation = System.getProperty(CIPHER_TRANSFORMATION_SYSTEM_PROPERTY);
+        if (StringUtils.isNotBlank(cipherTransformation) && returnSelfContainedCipherText) {
+            Certificate certificate = null;
+            try {
+                certificate = getCertificateFromStore();
+                encryptedKey = createSelfContainedCiphertext(encryptedKey, algorithm, certificate);
+            } catch (KeyStoreException | CertificateEncodingException | NoSuchAlgorithmException e) {
+                String errorMessage = String.format("An error occurred while decrypting using the algorithm : '%s'"
+                        , algorithm);
+
+                // Log the exception from client libraries, to avoid missing information if callers code doesn't log it
+                if (log.isDebugEnabled()) {
+                    log.debug(errorMessage, e);
+                }
+
+                throw new CryptoException(errorMessage, e);
+            }
+
+        }
+        return encryptedKey;
+    }
+
     private Certificate getCertificateFromStore() throws KeyStoreException {
 
         return keyStore.getCertificate(keyAlias);
@@ -171,5 +206,42 @@ public class KeyStoreBasedInternalCryptoProvider implements InternalCryptoProvid
         } else {
             return null;
         }
+    }
+
+    private byte[] createSelfContainedCiphertext(byte[] originalCipher, String transformation, Certificate certificate)
+            throws CertificateEncodingException, NoSuchAlgorithmException {
+
+        Gson gson = new Gson();
+        CipherMetaDataHolder cipherHolder = new CipherMetaDataHolder();
+        cipherHolder.setCipherText(Base64.encode(originalCipher));
+        cipherHolder.setTransformation(transformation);
+        cipherHolder.setThumbPrint(calculateThumbprint(certificate, "SHA-1"), "SHA-1");
+        String cipherWithMetadataStr = gson.toJson(cipherHolder);
+        if (log.isDebugEnabled()) {
+            log.debug("Cipher with meta data : " + cipherWithMetadataStr);
+        }
+        return cipherWithMetadataStr.getBytes(Charset.defaultCharset());
+    }
+
+    private String calculateThumbprint(Certificate certificate, String digest)
+            throws NoSuchAlgorithmException, CertificateEncodingException {
+
+        MessageDigest messageDigest = MessageDigest.getInstance(digest);
+        messageDigest.update(certificate.getEncoded());
+        byte[] digestByteArray = messageDigest.digest();
+        char[] hexCharacters = new char[]{'0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B',
+                'C', 'D', 'E', 'F'};
+
+
+        // convert digest in form of byte array to hex format
+        StringBuffer strBuffer = new StringBuffer();
+
+        for (int i = 0; i < digestByteArray.length; i++) {
+            int leftNibble = (digestByteArray[i] & 0xF0) >> 4;
+            int rightNibble = (digestByteArray[i] & 0x0F);
+            strBuffer.append(hexCharacters[leftNibble]).append(hexCharacters[rightNibble]);
+        }
+
+        return strBuffer.toString();
     }
 }
